# src/api-spec/ontology.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tellus Ontology Engine \u2014 Full API",
    "description": "Comprehensive API specification for the Tellus Ontology Engine. Covers ontology CRUD, object types, properties, link types, and action types.",
    "version": "2.0.0",
    "contact": {
      "name": "Tellus Engineering",
      "email": "engineering@tellus.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v2/ontologies/{ontologyId}/objectTypes": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OntologyId"
        }
      ],
      "post": {
        "operationId": "createObjectType",
        "summary": "Create a new object type",
        "description": "Creates a new object type within the specified ontology.",
        "tags": [
          "Object Types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "apiName",
                  "displayName"
                ],
                "properties": {
                  "apiName": {
                    "type": "string",
                    "description": "Unique programmatic name (e.g. 'customer', 'order')",
                    "example": "customer"
                  },
                  "displayName": {
                    "type": "string",
                    "description": "Human-readable name",
                    "example": "Customer"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional description"
                  },
                  "icon": {
                    "type": "string",
                    "nullable": true,
                    "description": "Icon identifier (e.g. an emoji or icon library key)"
                  },
                  "iconColor": {
                    "type": "string",
                    "nullable": true,
                    "description": "Icon color (hex or named color)"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "deprecated",
                      "experimental"
                    ],
                    "default": "active",
                    "description": "Lifecycle status of the object type"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Object type created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectType"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Object type with that apiName already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listObjectTypes",
        "summary": "List object types",
        "description": "Returns all object types within the specified ontology.",
        "tags": [
          "Object Types"
        ],
        "responses": {
          "200": {
            "description": "A list of object types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ObjectType"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/objectTypes/{apiName}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OntologyId"
        },
        {
          "$ref": "#/components/parameters/ObjectTypeApiName"
        }
      ],
      "get": {
        "operationId": "getObjectType",
        "summary": "Get a single object type",
        "description": "Returns the object type identified by apiName within the given ontology.",
        "tags": [
          "Object Types"
        ],
        "responses": {
          "200": {
            "description": "The requested object type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectType"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateObjectType",
        "summary": "Update an object type",
        "description": "Updates an existing object type's display name, description, icon, or status.",
        "tags": [
          "Object Types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "icon": {
                    "type": "string",
                    "nullable": true
                  },
                  "iconColor": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "deprecated",
                      "experimental"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Object type updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectType"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteObjectType",
        "summary": "Delete an object type",
        "description": "Permanently deletes the object type and all its properties.",
        "tags": [
          "Object Types"
        ],
        "responses": {
          "204": {
            "description": "Object type deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v2/objects/{objectType}/export": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export objects as CSV",
        "parameters": [
          {
            "name": "objectType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "csv"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "OntologyId": {
        "name": "ontologyId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID of the ontology"
      },
      "ObjectTypeApiName": {
        "name": "apiName",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Programmatic name of the object type"
      },
...
    },
    "schemas": {
...
      "ObjectType": {
        "type": "object",
        "properties": {
          "objectTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "ontologyId": {
            "type": "string",
            "format": "uuid"
          },
          "apiName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "iconColor": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "deprecated",
              "experimental"
            ]
          },
          "propertyCount": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "errorName": {
            "type": "string"
          },
          "errorInstanceId": {
            "type": "string"
          },
          "parameters": {
            "type": "object"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}